# Improve ktlint performances when the input patterns are absolute paths