## Description

This PR proposes a solution for the performance issue described in #1135.

When ktlint gets a bunch of absolute paths there is no need to walk through the file system and search for them, we can just resolve and analyze them directly.

Before:

```
➜  android-listeners git:(rc/kitlint) ktlint -F --debug --relative MyFile.kt
[DEBUG] Discovered ruleset with "standard" id.
[DEBUG] Discovered reporter with "baseline" id.
[DEBUG] Discovered reporter with "checkstyle" id.
[DEBUG] Discovered reporter with "json" id.
[DEBUG] Discovered reporter with "html" id.
[DEBUG] Discovered reporter with "plain" id.
[DEBUG] Initializing "plain" reporter with {verbose=true, color=false, color_name=DARK_GRAY}
[DEBUG] Checking MyFile.kt
Loaded .editorconfig: [insert_final_newline: true, max_line_length: 200]
[DEBUG] 30291ms / 1 file(s) / 0 error(s)
```

After:
```
➜  android-listeners git:(rc/kitlint) java -jar ktlint-0.42.0-SNAPSHOT-all.jar -F --debug --relative MyFile.kt
[DEBUG] Discovered ruleset with "standard" id.
[DEBUG] Discovered reporter with "baseline" id.
[DEBUG] Discovered reporter with "checkstyle" id.
[DEBUG] Discovered reporter with "json" id.
[DEBUG] Discovered reporter with "html" id.
[DEBUG] Discovered reporter with "plain" id.
[DEBUG] Initializing "plain" reporter with {verbose=true, color=false, color_name=DARK_GRAY}
[DEBUG] Checking MyFile.kt
Loaded .editorconfig: [insert_final_newline: true, max_line_length: 200]
[DEBUG] 847ms / 1 file(s) / 0 error(s)
```

In our project it speeds up the execution from ~30s down to less than 1s.

No additional tests were added, the exist ones passes.

## Checklist

<!-- Following checklist maybe skipped in some cases -->
- [ ] tests are added
- [ ] `CHANGELOG.md` is updated