@rciovati 2021-05-11
it looks like the build fails because of a dependency verification issue which I believe is not related to my change but I'll see what I can do.

```
For more details see https://docs.gradle.org/6.8.3/release-notes.html

Starting a Gradle Daemon (subsequent builds will be faster)
Dependency verification is an incubating feature.
> Task :buildSrc:extractPrecompiledScriptPluginPlugins
> Task :buildSrc:generateExternalPluginSpecBuilders FAILED

FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':buildSrc:generateExternalPluginSpecBuilders'.
> Dependency verification failed for configuration ':buildSrc:compileClasspath'
  2 artifacts failed verification:
    - dokka-gradle-plugin-0.10.1.jar (org.jetbrains.dokka:dokka-gradle-plugin:0.10.1) from repository MavenRepo
    - dokka-gradle-plugin-0.10.1.pom (org.jetbrains.dokka:dokka-gradle-plugin:0.10.1) from repository MavenRepo
  If the artifacts are trustworthy, you will need to update the gradle/verification-metadata.xml file by following the instructions at https://docs.gradle.org/6.8.3/userguide/dependency_verification.html#sec:troubleshooting-verification
  
  Open this report for more details: file:///home/runner/work/ktlint/ktlint/build/reports/dependency-verification/at-1620670437192/dependency-verification-report.html

* Try:
Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output. Run with --scan to get full insights.

* Get more help at https://help.gradle.org

BUILD FAILED in 49s
Error: Process completed with exit code 1.
```