@shashachu 2021-05-25
@rciovati do you have this commit rebased in? https://github.com/pinterest/ktlint/pull/1152/commits/fae23623a692c897f44cea91aed441dc60625921 

It might be the same issue.