@shashachu 2021-06-14
@rciovati @Tapchicoma `Path.resolve` seems to throw an exception on Windows when you pass in file globs. This change makes the tests pass; does it seem valid or too hacky?

```
val (existingFiles, actualGlobs) = globs.partition {
        try {
            Files.isRegularFile(rootDir.resolve(it))
        } catch (e: InvalidPathException) {
            false
        }
    }
```