@shashachu 2021-06-15
@rciovati looks like you are missing an import :) `e: /home/runner/work/ktlint/ktlint/ktlint/src/main/kotlin/com/pinterest/ktlint/internal/FileUtils.kt: (31, 21): Unresolved reference: InvalidPathException`